+++ packages/help-center/src/components/help-center-article-content.scss
@@ -1,7 +1,7 @@
 @import "@automattic/typography/styles/variables";
 @import "calypso/assets/stylesheets/shared/mixins/breakpoints";
 
-.help-center-article-content__story-content {
+.help-center-article-content__main {
 	color: var(--color-neutral-70);
 	font-size: $font-body;
 	line-height: 1.7;
