+++ packages/help-center/src/components/help-center-search.scss
@@ -377,7 +377,7 @@
  	 * ARTICLE EMBED
  	 */
 
-	.help-center-embed-result__header {
+	.help-center-article__header {
 		top: 0;
 		background: var(--color-surface);
 		position: sticky;
@@ -385,15 +385,15 @@
 		border-bottom: 1px solid var(--studio-gray-0);
 		padding: 0 10px;
 
-		.help-center-embed-result__external-button {
+		.help-center-article__external-button {
 			display: flex;
 			align-items: center;
 		}
 	}
-	.help-center-embed-result article.help-center-article-content__story {
+	.help-center-article article.help-center-article-content {
 		padding: 0 20px 20px;
 
-		h1.support-article-dialog__header-title > a {
+		h1.help-center-article-content__header-title > a {
 			text-decoration: none;
 			clear: none;
 			color: var(--color-neutral-70);
@@ -436,7 +436,7 @@
 			border-color: inherit;
 		}
 
-		a:not(.support-article-dialog__header-title-link) {
+		a:not(.help-center-article-content__header-title-link) {
 			text-decoration: none;
 		}
 
@@ -492,7 +492,7 @@
 			color: var(--studio-yellow-20);
 		}
 
-		.support-article-dialog__header {
+		.help-center-article-content__header {
 			margin-bottom: 0;
 		}
