+++ packages/help-center/src/hooks/index.ts
@@ -6,3 +6,9 @@ export { default as useChatStatus } from './use-chat-status';
 export { useShouldUseWapuu } from './use-should-use-wapuu';
 export { filterListBySearchTerm } from './use-admin-results';
 export { useActionHooks } from './use-action-hooks';
+export { usePostByKey } from './use-post-by-key';
+export {
+	useSupportArticleAlternatePostKey,
+	useSupportArticleAlternatesQuery,
+	getPostKey,
+} from './use-support-article-alternates-query';
